+++ tests/test_webserver/test_webserver_api_top_of_book.py
@@ -3,7 +3,7 @@ from fastapi import FastAPI
 from fastapi import status
 from fastapi.testclient import TestClient
 
-from limit_order_book_webserver.fastapi_webserver import app
+from limit_order_book_webserver.webserver import app
 from limit_order_book.types import OrderId
 from limit_order_book.exceptions import DuplicateOrderIdError
